### 2023-10-24 #5980 @editorialbot
[PRE REVIEW]: Pylinac: Image analysis for routine quality assurance in radiotherapy
**Submitting author:** @jrkerns (<a href="http://orcid.org/0000-0002-2019-7627">James Kerns</a>)
**Repository:** https://github.com/jrkerns/pylinac
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 3.16
