### 2023-10-25 bb764ac
use imagepositionpatient primarily
