+++ pylinac/core/image.py
@@ -1230,9 +1230,9 @@ class DicomImage(BaseImage):
     def z_position(self) -> float:
         """The z-position of the slice. Relevant for CT and MR images."""
         try:
-            return self.metadata.SliceLocation
-        except AttributeError:
             return self.metadata.ImagePositionPatient[-1]
+        except AttributeError:
+            return self.metadata.SliceLocation
 
     @property
     def slice_spacing(self) -> float:
