+++ docs/source/changelog.rst
@@ -13,6 +13,15 @@ Planar Imaging
   ROIs. This result will also be included in the ``results_data`` structure.
   This is not done for light/rad phantoms.
 
+ACR
+^^^
+
+* The ``z_position`` property for DICOM stacks (used in CT and MRI) was using ``SliceLocation`` if the tag
+  existed and ``ImagePositionPatient[-1]`` if it did not exist. The ``SliceLocation`` tag however is apparently
+  relative. This caused problems for the ACR MRI module on properly-acquired datasets.
+  The ``ImagePositionPatient`` tag is now the primary lookup key and ``SliceLocation`` is only used
+  if the former tag is unavailable.
+
 v 3.16.0
 --------
 
