### 2023-10-25 b4fb7e0
make docs strict
