+++ requirements-ram.txt
@@ -52,7 +52,7 @@ rich==12.4.4
 scikit-image==0.19.1
 scipy==1.7.3
 six==1.16.0
-sphinx
+sphinx==7.*
 sphinx-autobuild
 sphinx-copybutton
 tabulate~=0.9.0
