+++ bitbucket-pipelines.yml
@@ -21,7 +21,7 @@ definitions:
           - apt-get update
           - python -m venv venv
           - source venv/bin/activate
-          - pip install -r requirements-ram.txt
+          - pip install -r requirements-dev.txt
           - pip freeze
         artifacts:
           - venv/**
