### 2023-10-26 4578ff8
show couch and coll as disks; tweaks to visualization
