+++ bitbucket-pipelines.yml
@@ -247,6 +247,7 @@ definitions:
           - apt-get update
           - apt-get -y install curl
           - curl -sSfL https://raw.githubusercontent.com/anchore/grype/main/install.sh | sh -s -- -b /usr/local/bin
+          - pip freeze
           # we want to dump this to file regardless of fail status so we can inspect if needed
           - grype dir:. --only-fixed -o json > grype_results.json
           - grype dir:. --only-fixed --fail-on high
