### 2023-10-26 bb8b717
remove grype scanning
