+++ bitbucket-pipelines.yml
@@ -306,13 +306,6 @@ pipelines:
           - step: *vmat-tests
           - step: *winston-lutz-tests
           - step: *core-module-tests
-          - step:
-              <<: *grype-scanning
-              condition:
-                changesets:
-                  includePaths:
-                    - "requirements.txt"
-                    - "requirements-ram.txt"
 
   branches:
     release*:
