### 2023-10-27 b650421
add global finder + tests
