+++ tests_basic/core/test_image.py
@@ -552,11 +552,14 @@ class TestXIMImage(TestCase):
         assert_array_almost_equal(png_array, xim.array)
 
     def test_overflow(self):
+        old_settings = np.geterr()
         # set overflow to cause errors
         # shouldn't raise
         np.seterr(all="raise")
         xim_path = get_file_from_cloud_test_repo(["IsoCal-kV-08.xim"])
         XIM(xim_path)
+        # reset to old settings
+        np.seterr(**old_settings)
 
 
 class TestLinacDicomImage(TestCase):
