+++ requirements-ram.txt
@@ -35,7 +35,7 @@ pluggy==1.0.0
 py==1.11.0
 py-linq==1.3.0
 pydicom==2.2.2
-Pygments==2.12.0
+Pygments
 pyparsing==3.0.6
 pytest==7.1.2
 pytest-cov==3.0.0
@@ -52,7 +52,7 @@ rich==12.4.4
 scikit-image==0.19.1
 scipy==1.7.3
 six==1.16.0
-sphinx==5.*
+sphinx==7.*
 sphinx-autobuild
 sphinx-copybutton
 tabulate~=0.9.0
