+++ tests_basic/core/test_image.py
@@ -852,9 +852,8 @@ class TestGamma2D(TestCase):
 
 class TestTiffToDicom(TestCase):
     def test_conversion_can_be_loaded_as_dicom(self):
-        tiff_to_dicom(
+        ds = tiff_to_dicom(
             tif_path,
-            "output_dicom.dcm",
             sid=1000,
             dpi=200,
             gantry=10,
@@ -862,36 +861,32 @@ class TestTiffToDicom(TestCase):
             couch=33,
         )
         # shouldn't raise
-        LinacDicomImage("output_dicom.dcm")
+        LinacDicomImage.from_dataset(ds)
 
     def test_conversion_captures_axes(self):
-        tiff_to_dicom(
+        ds = tiff_to_dicom(
             tif_path,
-            "output_dicom.dcm",
             sid=1000,
             dpi=200,
             gantry=10,
             coll=22,
             couch=33,
         )
-        dicom_img = LinacDicomImage("output_dicom.dcm")
+        dicom_img = LinacDicomImage.from_dataset(ds)
         self.assertEqual(dicom_img.gantry_angle, 10)
         self.assertEqual(dicom_img.collimator_angle, 22)
         self.assertEqual(dicom_img.couch_angle, 33)
 
     def test_conversion_of_dpmm(self):
-        tiff_to_dicom(
-            tif_path, "output_dicom.dcm", sid=1000, gantry=10, coll=22, couch=33
-        )
-        dicom_img = LinacDicomImage("output_dicom.dcm")
+        ds = tiff_to_dicom(tif_path, sid=1000, gantry=10, coll=22, couch=33)
+        dicom_img = LinacDicomImage.from_dataset(ds)
         self.assertEqual(dicom_img.dpi, 150)
         self.assertEqual(dicom_img.dpmm, 150 / 25.4)
 
     def test_conversion_goes_to_uint16(self):
         tiff_img = FileImage(tif_path)
-        tiff_to_dicom(
+        ds = tiff_to_dicom(
             tif_path,
-            "output_dicom.dcm",
             sid=1000,
             dpi=200,
             gantry=10,
@@ -899,23 +894,12 @@ class TestTiffToDicom(TestCase):
             couch=33,
         )
         self.assertEqual(tiff_img.array.dtype, np.uint8)
-        dicom_img = LinacDicomImage("output_dicom.dcm")
+        dicom_img = LinacDicomImage.from_dataset(ds)
         self.assertEqual(dicom_img.array.dtype, np.uint16)
-
-    def test_conversion_to_stream(self):
-        with io.BytesIO() as stream:
-            tiff_to_dicom(
-                tif_path, stream, sid=1000, dpi=200, gantry=0, coll=0, couch=0
-            )
-            dicom_img = LinacDicomImage(stream)
-        self.assertEqual(dicom_img.gantry_angle, 0)
 
     def test_mass_conversion(self):
         """Mass conversion; shouldn't fail. All images have dpi tag"""
         all_starshot_files = get_folder_from_cloud_test_repo(["Starshot"])
         for img in Path(all_starshot_files).iterdir():
             if img.suffix in (".tif", ".tiff"):
-                with io.BytesIO() as stream:
-                    tiff_to_dicom(
-                        img, dicom_file=stream, sid=1000, gantry=10, coll=11, couch=12
-                    )
+                tiff_to_dicom(img, sid=1000, gantry=10, coll=11, couch=12)
