### 2023-10-30 939ca3d
add x-values checks
