### 2023-10-30 #5980 @ProfLeao
Yes, I can!
