### 2023-10-31 jrkerns/pylinac#467 @SimonBiggs
Doc badge link in readme broken 
Hi James ☺️

The doc badge link within the readme is broken. Click on the badge for the docs and you get a 404
