### 2023-11-01 jrkerns/pylinac#467 @jrkerns
Haha, wow that must've been broken a long time. I'll have a fix up soon. Thanks!
