### 2023-11-01 9272f0f
fix broken link from GH page Closes #467
