+++ bitbucket-pipelines.yml
@@ -42,6 +42,7 @@ definitions:
           changesets:
             includePaths:
               - "**/*.rst"
+              - "*.rst"
     - step: &cbct-tests
         name: Run CBCT Tests
         script:
