### 2023-11-01 afbe046
Merged in bugfix/docs_badge (pull request #288)
fix broken link from GH page Closes #467

* fix broken link from GH page Closes #467

* change to trigger on any rst change

* asterisk

* unsure
