### 2023-11-02 jrkerns/pylinac#467 @SimonBiggs
My pleasure 🙂. I find these JOSS reviews help find a few little things that tend to hang around the edges. It's quite a nice process. 🙂
