### 2023-11-02 jrkerns/pylinac#468 @SimonBiggs
Contributing guide
Hi @jrkerns,

One thing that might be helpful is expanding the contributor part of the README to also a, "and if you do happen to be a code contributor here is how to set up a dev environment and submit a PR. I'd be more than happy to help in this process if you want to reach out":

https://github.com/jrkerns/pylinac#contributing

Maybe most of the bones of this can be in here:
https://pylinac.readthedocs.io/en/latest/contributing.html#commit-changes

And then linked to in the README?
