### 2023-11-02 #6001 @ProfLeao
Hi @jrkerns, I believe you've already seen it, but several links in your main README file are broken, However, this has no impacto on the package functionality  nor on the revision level required by JOSS. If you like, I can mark them for easy identification.
