### 2023-11-02 #6001 @ProfLeao
Just for record, I would like to note that no explicit "community guidelines" were found. However, I considered the code highlighted below to be sufficient.

https://github.com/jrkerns/pylinac/blob/c6286ca6c86e90a995a0cad674572991fcc5d30e/docs/source/overview.rst?plain=1#L46-L104
