### 2023-11-02 #6001 @kellyrowland
Thanks; I might suggest adding instructions on how to contribute to the project in this section (i.e., are folks welcome to open issues/PRs that follow in this philosophy?).
