### 2023-11-04 #6001 @SimonBiggs
Also, if it helps @ProfLeao, I made the following issue here:

https://github.com/jrkerns/pylinac/issues/468#issue-1973303026
