### 2023-11-06 #6001 @ProfLeao
> Also, if it helps @ProfLeao, I made the following issue here:
>
> [jrkerns/pylinac#468 (comment)](https://github.com/jrkerns/pylinac/issues/468#issue-1973303026)

Great! Thanks.
