### 2023-11-06 #6001 @jrkerns
Hi @ProfLeao and @SimonBiggs; thanks for taking the time to review the submission. Simon created two issues which I recently addressed: https://github.com/jrkerns/pylinac/issues/468 (contribution guide) and https://github.com/jrkerns/pylinac/issues/467 (RTD badge link). @ProfLeao I also [fixed the links](https://github.com/jrkerns/pylinac/commit/8b64043af46c426097db6fd96b471a45018ddd06) in the headline sections from `stable` to `latest` which was causing broken links. These are fixed and pushed already so should be live. I poked around and didn't find any other broken links, but if there are any others you had run into please let me know. @kellyrowland I added a bit to the [contributing section](https://pylinac.readthedocs.io/en/latest/contributing.html). If this needs further elaboration please let me know. Thanks all!
