### 2023-11-07 #6001 @ProfLeao
Hi @jrkerns , I'll leave some "suggestions" for improving the text. Please note that my native language is not English, so any suggestions that are not relevant can be disregarded.

## Summary

### 1. **Line:** *6*

Original text: "Pylinac is a python library (...)"

Suggested text: "Pylinac is a Python library (...)"

### 2. **Line:** *9*

Original text: "required by the medical physics society."

Suggested text: "required by the Medical Physics Society."

### 3. **Line:** *10*

Original text: as the target audience is not developers. Thus, most workflows
can be executed in a few lines (...)"

Suggested text: "as the target audience is not developers. Thus, most
workflows can be implemented in a few lines of code."

## Statement of need

### 4. **Line:** *23*

Original text: "multileaf collimator that shape the radiation
(Calvo-Ortega et al., 2014) These images and (...)"

Suggested text: "multileaf collimator that shape the radiation
(Calvo-Ortega et al., 2014). These images and (...)"

## Example usage

### 5. **Line:** *56*

Original text: "where 𝐼 is the ROI of the contrast region in question (...)"

Suggested text: "where 𝐼 is the ROI of the contrast region of interest (...)"

### 6. **Line:** *58*

Original text: "This corresponds to the circular ROIs on the outer edge (...)"

Suggested text: "This corresponds to the circular ROIs at the outer edge (...)"

### 7. **Line:** *69*

Original text: "69 be saved as records that may be audited by state
authorities. (...)"

Suggested text: "69 be stored as records that can be audited by government
authorities. (...)"
