### 2023-11-07 416608f
Fix non-integer x-values for profiles
