+++ pylinac/core/profile.py
@@ -272,24 +272,26 @@ class SymmetryPointDifferenceMetric(ProfileMetric):
         max_sym_idx = np.argmax(np.abs(self.symmetry_values))
         return self.symmetry_values[max_sym_idx]
 
-    def plot(self, axis: plt.Axes) -> None:
+    def plot(self, axis: plt.Axes, markers: (str, str) = ("^", "v")) -> None:
         idx = np.argmax(self.symmetry_values)
         left_edge, right_edge, _ = self.profile.field_indices(
             in_field_ratio=self.in_field_ratio
         )
         # plot max sym value
+        max_x = self.profile.x_at_x_idx(self.profile.x_idx_at_x(left_edge) + idx)
         axis.plot(
-            left_edge + idx,
-            self.profile.y_at_x(left_edge + idx),
-            "^",
+            max_x,
+            self.profile.y_at_x(max_x),
+            markers[0],
             color=self.color,
             label=self.name,
         )
         # plot min sym value
+        min_x = self.profile.x_at_x_idx(self.profile.x_idx_at_x(right_edge) - idx)
         axis.plot(
-            right_edge - idx,
-            self.profile.y_at_x(right_edge - idx),
-            "v",
+            min_x,
+            self.profile.y_at_x(min_x),
+            markers[1],
             color=self.color,
         )
 
@@ -303,7 +305,7 @@ class SymmetryPointDifferenceMetric(ProfileMetric):
         ylim_bottom = min((min(self.symmetry_values) - 0.5, self.min_sym - 0.5))
         sec_ax.set_ylim(ylim_bottom, ylim_top)
         sec_ax.plot(
-            range(left_edge, len(self.symmetry_values) + left_edge),
+            self.profile.field_x_values(self.in_field_ratio),
             self.symmetry_values,
             color=self.color,
             linestyle=self.linestyle,
@@ -330,6 +332,9 @@ class SymmetryPointDifferenceQuotientMetric(SymmetryPointDifferenceMetric):
         """Calculate an individual point's symmetry."""
         return 100 * max((lt / rt), (rt / lt))
 
+    def plot(self, axis: plt.Axes, markers: (str, str) = ("x", "x")) -> None:
+        super().plot(axis, markers)
+
 
 class SymmetryAreaMetric(ProfileMetric):
     """The symmetry using ratios of the areas of the left and right sides of the profile."""
@@ -355,15 +360,15 @@ class SymmetryAreaMetric(ProfileMetric):
 
     def plot(self, axis: plt.Axes):
         """Plot the symmetry by shading the left and right areas"""
-        left, right, width = self.profile.field_indices(
-            in_field_ratio=self.in_field_ratio
-        )
-        left_data = self.profile.field_values(self.in_field_ratio)[: width // 2 + 1]
-        left_stretch = np.linspace(left, left + width // 2, len(left_data))
-        right_data = self.profile.field_values(self.in_field_ratio)[width // 2 :]
-        right_stretch = np.linspace(right - width // 2, right, len(right_data))
-        axis.fill_between(left_stretch, left_data, alpha=0.2, label="Left Area")
-        axis.fill_between(right_stretch, right_data, alpha=0.2, label="Right Area")
+        field_values = self.profile.field_values(self.in_field_ratio)
+        x_values = self.profile.field_x_values(self.in_field_ratio)
+        split = math.floor(len(field_values) / 2)
+        left_data = field_values[: split + 1]
+        left_x = x_values[: split + 1]
+        right_data = field_values[split:]
+        right_x = x_values[split:]
+        axis.fill_between(left_x, left_data, alpha=0.2, label="Left Area")
+        axis.fill_between(right_x, right_data, alpha=0.2, label="Right Area")
 
 
[…]
