+++ pylinac/core/metrics.py
@@ -76,9 +76,11 @@ def is_right_circumference(region: RegionProperties, *args, **kwargs) -> bool:
 
 def is_right_square_perimeter(region: RegionProperties, *args, **kwargs) -> bool:
     """Test the regionprop's perimeter attr to see if it matches
-    that of an equivalent square"""
+    that of an equivalent square. In reality, edges aren't perfectly straight, so
+    the real perimeter is always going to be higher than the theoretical perimeter.
+    We thus add a larger tolerance (20%) to the upper perimeter"""
     actual_perimeter = region.perimeter / kwargs["dpmm"]
-    upper_perimeter = 2 * (
+    upper_perimeter = 1.20 * 2 * (
         kwargs["field_width_mm"] + kwargs["field_tolerance_mm"]
     ) + 2 * (kwargs["field_height_mm"] + kwargs["field_tolerance_mm"])
     lower_perimeter = 2 * (
@@ -496,37 +498,158 @@ class GlobalDiskLocator(MetricBase):
 
 class GlobalSizedFieldLocator(MetricBase):
     fields: list[Point]
+    boundaries: list[np.ndarray]
+    is_from_physical: bool = False
 
     def __init__(
         self,
-        field_with_mm: float,
-        field_height_mm: float,
-        field_tolerance_mm: float,
+        field_width_px: float,
+        field_height_px: float,
+        field_tolerance_px: float,
         min_number: int = 1,
         max_number: int | None = None,
-        min_separation_mm: float = 5,
         name: str = "Field Finder",
         detection_conditions: list[callable] = (
             is_right_square_perimeter,
             is_right_area_square,
         ),
+        default_threshold_step_size: float = 2,
     ):
-        self.field_width_mm = field_with_mm
-        self.field_height_mm = field_height_mm
-        self.field_tolerance_mm = field_tolerance_mm
+        """Finds fields globally within an image.
+
+        Parameters
+        ----------
+        field_width_px : float
+            The width of the field in px.
+        field_height_px : float
+            The height of the field in px.
+        field_tolerance_px : float
+            The tolerance of the field size in px.
+        min_number : int
+            The minimum number of fields to find. If not found, an error is raised.
+        max_number : int, None
+            The maximum number of fields to find. If None, no maximum is set.
+        name : str
+            The name of the metric.
+        detection_conditions : list[callable]
+            A list of functions that take a regionprops object and return a boolean.
+        default_threshold_step_size : float
+            The default step size for the threshold iteration. This is based on the max number of fields and the field size.
+        """
+        self.field_width_mm = field_width_px
+        self.field_height_mm = field_height_px
+        self.field_tolerance_mm = field_tolerance_px
         self.min_number = min_number
         self.max_number = max_number or 1e6
         self.name = name
         self.detection_conditions = detection_conditions
-        self.min_separation_mm = min_separation_mm
+        self.default_threshold_step_size = default_threshold_step_size
+
+    @classmethod
+    def from_physical(
+        cls,
+        field_width_mm: float,
+        field_height_mm: float,
+        field_tolerance_mm: float,
+        min_number: int = 1,
+        max_number: int | None = None,
+        name: str = "Field Finder",
+        detection_conditions: list[callable] = (
+            is_right_square_perimeter,
+            is_right_area_square,
+        ),
+        default_threshold_step_size: float = 2,
+    ):
+        """Construct an instance using physical dimensions.
+
+        Parameters
+        ----------
+        field_width_mm : float
+            The width of the field in mm.
+        field_height_mm : float
+            The height of the field in mm.
+        field_tolerance_mm : float
+            The tolerance of the field size in mm.
+        min_number : int
+            The minimum number of fields to find. If not found, an error is raised.
+        max_number : int, None
[…]
