+++ joss/paper.md
@@ -19,51 +19,51 @@ bibliography:
 
 # Summary
 
-Pylinac is a python library that analyzes routine quality assurance images generated by 
-devices in the radiotherapy treatment domain. It contains multiple independent modules 
-that map to different quality assurance tasks and will ingest and analyze the images for common metrics as required by the medical 
-physics society. The library is designed to be concise and easy to use as the target audience is not 
-developers. Thus, most workflows can be executed in a few lines of code. At the same time,
-the library is modular and allows for easy extension via plugins and configuration settings. 
+Pylinac is a Python library that analyzes routine quality assurance images generated by
+devices in the radiotherapy treatment domain. It contains multiple independent modules
+that map to different quality assurance tasks and will ingest and analyze the images for common metrics as required by the medical
+physics society. The library is designed to be concise and easy to use as the target audience is not
+developers. Thus, most workflows can be implemented in a few lines of code. At the same time,
+the library is modular and allows for easy extension via plugins and configuration settings.
 Comprehensive documentation is available with usage examples and algorithm explanations.
 
 
 # Statement of need
 
-Within the therapeutic medical physics domain, verification that radiotherapy machines are 
-performing in accordance with specification is an obvious need considering their use on 
-humans and is required by relevant societies [@Kutcher1994; @Klein2009; @Dieterich2008]. This involves 
-routine quality assurance (QA) at regular intervals by medical physicists. A subset of this QA 
-involves acquisition and analysis of images generated by the radiotherapy devices. 
-This includes the mechanical size of the "isocenter" of the linear accelerator [@Winston1988], 
+Within the therapeutic medical physics domain, verification that radiotherapy machines are
+performing in accordance with specification is an obvious need considering their use on
+humans and is required by relevant societies [@Kutcher1994; @Klein2009; @Dieterich2008]. This involves
+routine quality assurance (QA) at regular intervals by medical physicists. A subset of this QA
+involves acquisition and analysis of images generated by the radiotherapy devices.
+This includes the mechanical size of the "isocenter" of the linear accelerator [@Winston1988],
 dosimetric performance of the accelerator as it rotates around the patient [@Ling2008], and examination
-of the individual "leaves" of the multileaf collimator that shape the radiation [@Calvo2014]
+of the individual "leaves" of the multileaf collimator that shape the radiation [@Calvo2014].
 These images and data test the various mechanical and dosimetric performance dimensions of the machine.
 The images are usually the same pattern at every interval and are used for constancy testing.
 Manual examination of images is subject to interpersonal interpretation [@Ho1995; @Kerns2013].
-Performing this quality assurance has been examined as being quantifiable by image or digital analysis in the past [@Depuydt2012; @Rowshanfarzad2011; @Du2009; @Eckhause2015; @Jorgensen2011; @Kerns2014]. 
+Performing this quality assurance has been examined as being quantifiable by image or digital analysis in the past [@Depuydt2012; @Rowshanfarzad2011; @Du2009; @Eckhause2015; @Jorgensen2011; @Kerns2014].
 Commercial applications exist but can be prohibitively expensive
 and at the time the library was written no open-source alternatives existed.
-Medical physicists usually do not have computer science training and creating their 
-own in-house software for such evaluation can be difficult to justify. 
+Medical physicists usually do not have computer science training and creating their
+own in-house software for such evaluation can be difficult to justify.
 There is thus a need for software for budget-constrained radiotherapy clinics as well
[…]
