### 2023-11-09 #6001 @jrkerns
@ProfLeao Thank you for the suggestions. I implemented all of them except for the Summary Line 9 about medical physics society. In the US, there is a society called the [AAPM](https://w3.aapm.org/my_aapm/index.php), and there are several other "societies" internationally (e.g. [EFOMP](https://www.efomp.org/)). I didn't think it would be helpful to list out the specific societies, so I purposely used an informal term here for the community-as-a-whole. If this is acceptable, then I believe everything was addressed. Thank you for taking the time to review the submission. The updated proof is above and the [commit itself](https://github.com/jrkerns/pylinac/commit/2d22c37eed770c27bfb62635a4c0859128cd53ec). Thank you for taking the time to review the submission.
