### 2023-11-09 7638470
fix Halycon MLC config
