### 2023-11-09 46c42dc
add a hypersight quart variant
