### 2023-11-10 3b99bfc
drop demo file usage for starshot
