+++ pylinac/core/image.py
@@ -40,6 +40,7 @@ from .array_utils import (
     invert,
     normalize,
 )
+from .decorators import lru_cache
 from .geometry import Point
 from .io import (
     TemporaryZipDirectory,
@@ -1096,15 +1097,6 @@ class XIM(BaseImage):
             )
         return 1 / (10 * self.properties["PixelHeight"])
 
-    @property
-    def dpmm(self) -> float:
-        """The dots/mm value of the XIM images. The value appears to be in cm in the file."""
-        if self.properties["PixelWidth"] != self.properties["PixelHeight"]:
-            raise ValueError(
-                "The XIM image does not have the same pixel height and width"
-            )
-        return 1 / (10 * self.properties["PixelHeight"])
-
     def save_as(self, file: str, format: str | None = None) -> None:
         """Save the image to a NORMAL format. PNG is highly suggested. Accepts any format supported by Pillow.
         Ironically, an equivalent PNG image (w/ metadata) is ~50% smaller than an .xim image.
@@ -1552,7 +1544,113 @@ class ArrayImage(BaseImage):
         return ArrayImage(self.array - other.array)
 
 
-class DicomImageStack:
+class LazyDicomImageStack:
+    _image_path_keys: list[Path]
+
+    def __init__(
+        self,
+        folder: str | Path,
+        dtype: np.dtype | None = None,
+        min_number: int = 39,
+        check_uid: bool = True,
+    ):
+        """Load a folder with DICOM CT images. This variant is more memory efficient than the standard DicomImageStack.
+
+        This is done by loading images from disk on the fly. This assumes all images remain on disk for the lifetime of the instance. This does not
+        need to be true for the original implementation.
+
+        See the documentation for DicomImageStack for parameter descriptions.
+        """
+        self.dtype = dtype
+        paths = []
+        # load in images in their received order
+        if isinstance(folder, (list, tuple)):
+            paths = folder
+        elif osp.isdir(folder):
+            for pdir, sdir, files in os.walk(folder):
+                for file in files:
+                    paths.append(osp.join(pdir, file))
+        # we only want to read the metadata once
+        # so we read it here and then filter and sort
+        metadatas, paths = self._get_path_metadatas(paths)
+
+        # check that at least 1 image was loaded
+        if len(paths) < 1:
+            raise FileNotFoundError(
+                f"No files were found in the specified location: {folder}"
+            )
+
+        # error checking
+        if check_uid:
+            most_common_uid = self._get_common_uid_imgs(metadatas, min_number)
+            metadatas = [m for m in metadatas if m.SeriesInstanceUID == most_common_uid]
+        # sort according to physical order
+        order = np.argsort([m.ImagePositionPatient[-1] for m in metadatas])
+        self._image_path_keys = [paths[i] for i in order]
+
+    @classmethod
+    def from_zip(cls, zip_path: str | Path, dtype: np.dtype | None = None):
+        """Load a DICOM ZIP archive.
+
+        Parameters
+        ----------
+        zip_path : str
+            Path to the ZIP archive.
+        dtype : dtype, None, optional
+            The data type to cast the image data as. If None, will use whatever raw image format is.
+        """
+        with TemporaryZipDirectory(zip_path, delete=False) as tmpzip:
+            obj = cls(tmpzip, dtype)
+        return obj
+
+    def _get_common_uid_imgs(
+        self, metadata: list[pydicom.Dataset], min_number: int
+    ) -> pydicom.DataElement:
+        """Check that all the images are from the same study."""
+        most_common_uid = Counter(i.SeriesInstanceUID for i in metadata).most_common(1)[
+            0
+        ]
+        if most_common_uid[1] < min_number:
+            raise ValueError(
[…]
