+++ docs/source/conf.py
@@ -74,9 +74,9 @@ author = "James Kerns"
 # built documents.
 #
 # The short X.Y version.
-version = "3.16"
+version = "3.17"
 # The full version, including alpha/beta/rc tags.
-release = "3.16.0"
+release = "3.17.0"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
