### 2023-11-14 958c39c
patch for homogenous ROI
