### 2023-11-15 0bb08b4
drop grype and bump base pipe to 3.11
