+++ pyproject.toml
@@ -56,6 +56,7 @@ build-backend = "poetry.core.masonry.api"
 [tool.pytest.ini_options]
 testpaths = "tests_basic"
 python_files = "test_*.py"
+addopts = "--dist loadscope"
 
 [tool.coverage.run]
 source = [
