### 2023-11-16 #6001 @jrkerns
V 3.17
GH release: https://github.com/jrkerns/pylinac/releases/tag/v3.17
pypi: https://pypi.org/project/pylinac/3.17.0/
zenodo: https://zenodo.org/records/10145069
