### 2023-11-20 ad79dc5
rework the disk locator to match WL. Use locator in WL standard version.
