+++ pylinac/winston_lutz.py
@@ -53,6 +53,14 @@ from .core.geometry import Line, Point, Vector, cos, sin
 from .core.image import DicomImageStack, LinacDicomImage, is_image, tiff_to_dicom
 from .core.io import TemporaryZipDirectory, get_url, retrieve_demo_file
 from .core.mask import bounding_box
+from .core.metrics import (
+    DiskLocator,
+    is_right_circumference,
+    is_right_size_bb,
+    is_round,
+    is_solid,
+    is_symmetric,
+)
 from .core.scale import MachineScale, convert
 from .core.utilities import ResultBase, convert_to_enum, is_close
 
@@ -295,16 +303,6 @@ def plot_image(img: WinstonLutz2D | None, axis: plt.Axes) -> None:
         img.plot(ax=axis, show=False)
 
 
-def is_symmetric(region: RegionProperties, *args, **kwargs) -> bool:
-    """Whether the binary object's dimensions are symmetric, i.e. a perfect circle. Used to find the BB."""
-    ymin, xmin, ymax, xmax = region.bbox
-    y = abs(ymax - ymin)
-    x = abs(xmax - xmin)
-    if x > max(y * 1.05, y + 3) or x < min(y * 0.95, y - 3):
-        return False
-    return True
-
-
 def is_near_center(region: RegionProperties, *args, **kwargs) -> bool:
     """Whether the bb is <2cm from the center of the field"""
     dpmm = kwargs["dpmm"]
@@ -333,13 +331,6 @@ def is_modest_size(region: RegionProperties, *args, **kwargs) -> bool:
     return smaller_bb_area < bb_area < larger_bb_area
 
 
-def is_round(region: RegionProperties, *args, **kwargs) -> bool:
-    """Decide if the ROI is circular in nature by testing the filled area vs bounding box. Used to find the BB."""
-    expected_fill_ratio = np.pi / 4  # area of a circle inside a square
-    actual_fill_ratio = region.filled_area / region.bbox_area
-    return expected_fill_ratio * 1.2 > actual_fill_ratio > expected_fill_ratio * 0.8
-
-
 def is_square(region: RegionProperties, *args, **kwargs) -> bool:
     """Decide if the ROI is square in nature by testing the filled area vs bounding box. Used to find the BB."""
     actual_fill_ratio = region.filled_area / region.bbox_area
@@ -361,11 +352,13 @@ class WinstonLutz2D(image.LinacDicomImage):
     bb: Point
     field_cax: Point
     _rad_field_bounding_box: list
+    is_from_tiff: bool = False
     detection_conditions: list[callable] = [
+        is_right_size_bb,
         is_round,
+        is_right_circumference,
         is_symmetric,
-        is_near_center,
-        is_modest_size,
+        is_solid,
     ]
 
     def __init__(
@@ -393,15 +386,16 @@ class WinstonLutz2D(image.LinacDicomImage):
         bb_size_mm: float = 5,
         low_density_bb: bool = False,
         open_field: bool = False,
+        bb_tolerance_mm: float = 2,
     ) -> None:
         """Analyze the image. See WinstonLutz.analyze for parameter details."""
         self.field_cax, self._rad_field_bounding_box = self._find_field_centroid(
             open_field
         )
         if low_density_bb:
-            self.bb = self._find_low_density_bb(bb_size_mm)
+            self.bb = self._find_low_density_bb(bb_size_mm, bb_tolerance_mm)
         else:
-            self.bb = self._find_bb(bb_size_mm)
+            self.bb = self._find_bb(bb_size_mm, bb_tolerance_mm)
         self._is_analyzed = True
 
     def __repr__(self):
[…]

[…]
