### 2023-11-21 #6001 @Kevin-Mattheus-Moerman
@jrkerns As AEiC I will now help to process this work for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, however I do have the below points that need your attention:

- [ ] Your archive currently says `3.17.0` for the version, however this should be edited to match `v3.17`, i.e. drop the `0` and include the `v` so it will say end up listing it as `Version v3.17`.
- [ ] Please edit your affiliation by adding the town, and state as well. Also, please spell out the USA as `United States of America`.
- [ ] Check (I am unsure) if this sentence `Conceptualization, coding, development and paper writing by James Kerns.` needs an Oxford comma, i.e. `Conceptualization, coding, development, and paper writing by James Kerns.`.
- [ ] Recommendation (no action required): The acronym ROI is defined as `regions of interest (ROI)`, i.e. using the plural "regions". Later though it is both used as ROI and ROIs. Would it be less confusing to rewrite it to introduce it in a singular form e.g. `region of interest ROI`, and then to add the `s` when you want the plural form?
