+++ pylinac/winston_lutz.py
@@ -386,16 +386,15 @@ class WinstonLutz2D(image.LinacDicomImage):
         bb_size_mm: float = 5,
         low_density_bb: bool = False,
         open_field: bool = False,
-        bb_tolerance_mm: float = 2,
     ) -> None:
         """Analyze the image. See WinstonLutz.analyze for parameter details."""
         self.field_cax, self._rad_field_bounding_box = self._find_field_centroid(
             open_field
         )
         if low_density_bb:
-            self.bb = self._find_low_density_bb(bb_size_mm, bb_tolerance_mm)
+            self.bb = self._find_low_density_bb(bb_size_mm)
         else:
-            self.bb = self._find_bb(bb_size_mm, bb_tolerance_mm)
+            self.bb = self._find_bb(bb_size_mm)
         self._is_analyzed = True
 
     def __repr__(self):
@@ -465,68 +464,32 @@ class WinstonLutz2D(image.LinacDicomImage):
             p = Point(x=coords[-1], y=coords[0])
         return p, edges
 
-    def _find_low_density_bb(self, bb_diameter: float, bb_tolerance_mm: float) -> Point:
+    def _calculate_bb_tolerance(self, bb_diameter: float) -> int:
+        """Calculate the BB tolerance based on the BB diameter.
+        Min will be 2 for 1.5mm and under. Will be 4 for diameters at or above 30mm."""
+        y = (2, 4)
+        x = (1.5, 30)
+        return np.interp(bb_diameter, x, y)
+
+    def _find_low_density_bb(self, bb_diameter: float) -> Point:
         """Find the BB within the radiation field, where the BB is low-density and creates
         an *increase* in signal vs a decrease/attenuation. The algorithm is similar to the
         normal _find_bb, but there would be so many if-statements it would be very convoluted and contain superfluous variables
         """
+        bb_tolerance_mm = self._calculate_bb_tolerance(bb_diameter)
         center = self.compute(
             metrics=DiskLocator.from_center_physical(
                 expected_position_mm=(0, 0),
-                search_window_mm=(30 + bb_diameter, 30 + bb_diameter),
+                search_window_mm=(20 + bb_diameter, 20 + bb_diameter),
                 radius_mm=bb_diameter / 2,
-                radius_tolerance_mm=bb_tolerance_mm / 2,
+                radius_tolerance_mm=bb_tolerance_mm,
                 invert=False,
                 detection_conditions=self.detection_conditions,
             )
         )
         return center
-        # get initial starting conditions
-        lower_thresh = self.array.max() * 0.8
-        spread = self.array.max() - lower_thresh
-        # search for the BB by iteratively increasing a high-pass threshold value until the BB is found.
-        found = False
-        while not found:
-            try:
-                binary_arr = self >= lower_thresh
-                # use below for debugging
-                # plt.imshow(binary_arr)
-                # plt.show()
-                labeled_arr, num_roi = ndimage.label(binary_arr)
-                regions = measure.regionprops(labeled_arr)
-                conditions_met = [
-                    all(
-                        condition(
-                            region,
-                            dpmm=self.dpmm,
-                            bb_size=bb_diameter / 2,
-                            tolerance=bb_tolerance_mm / 2,
-                            shape=binary_arr.shape,
-                        )
-                        for condition in self.detection_conditions
-                    )
-                    for region in regions
-                ]
-                if not any(conditions_met):
-                    raise ValueError
-                else:
-                    region_idx = [
-                        idx for idx, value in enumerate(conditions_met) if value
-                    ][0]
-                    found = True
-            except (IndexError, ValueError):
[…]
