+++ tests_basic/test_winstonlutz.py
@@ -11,7 +11,6 @@ from unittest import TestCase
 import matplotlib.pyplot as plt
 import numpy as np
 
-import pylinac
 from pylinac import WinstonLutz, WinstonLutzMultiTargetMultiField
 from pylinac.core.array_utils import create_dicom_files_from_3d_array
 from pylinac.core.geometry import Vector, vector_is_close
@@ -523,7 +522,7 @@ class TestWLLoading(TestCase, FromDemoImageTesterMixin, FromURLTesterMixin):
                 r"AQA_B_03082023.tif": (0, 0, 0),
             },
         )
-        ref_w.analyze(bb_size_mm=24, bb_tolerance_mm=4)
+        ref_w.analyze(bb_size_mm=24)
         results = ref_w.results_data()
         self.assertEqual(results.num_gantry_images, 2)
 
@@ -587,11 +586,6 @@ class GeneralTests(TestCase):
 
         self.wl.plot_axis_images("Gantry")
         self.wl.plot_axis_images(Axis.GANTRY)
-
-    def test_bb_override(self):
-        with self.assertRaises(ValueError):
-            wl = pylinac.WinstonLutz.from_demo_images()
-            wl.analyze(bb_size_mm=8)
 
     def test_bb_shift_instructions(self):
         move = self.wl.bb_shift_instructions()
@@ -725,7 +719,6 @@ class WinstonLutzMixin(CloudFileMixin):
     num_images = 0
     zip = True
     bb_size = 5
-    bb_tolerance_mm = 2
     sid: float | None = None
     dpi: float | None = None
     axis_mapping: dict | None = None
@@ -747,10 +740,10 @@ class WinstonLutzMixin(CloudFileMixin):
     use_filenames = False
 
     @classmethod
-    def setUpClass(cls):
+    def new_instance(cls) -> WinstonLutz:
         filename = cls.get_filename()
         if cls.zip:
-            cls.wl = WinstonLutz.from_zip(
+            wl = WinstonLutz.from_zip(
                 filename,
                 use_filenames=cls.use_filenames,
                 sid=cls.sid,
@@ -758,23 +751,31 @@ class WinstonLutzMixin(CloudFileMixin):
                 axis_mapping=cls.axis_mapping,
             )
         else:
-            cls.wl = WinstonLutz(
+            wl = WinstonLutz(
                 filename,
                 use_filenames=cls.use_filenames,
                 sid=cls.sid,
                 dpi=cls.dpi,
                 axis_mapping=cls.axis_mapping,
             )
+        return wl
+
+    @classmethod
+    def setUpClass(cls):
+        cls.wl = cls.new_instance()
         cls.wl.analyze(
             bb_size_mm=cls.bb_size,
             machine_scale=cls.machine_scale,
             low_density_bb=cls.low_density_bb,
             open_field=cls.open_field,
-            bb_tolerance_mm=cls.bb_tolerance_mm,
         )
         if cls.print_results:
             print(cls.wl.results())
             print(cls.wl.bb_shift_vector)
+
+    @classmethod
+    def tearDownClass(cls):
+        plt.close("all")
 
     def test_number_of_images(self):
         self.assertEqual(self.num_images, len(self.wl.images))
@@ -835,10 +836,33 @@ class WinstonLutzMixin(CloudFileMixin):
             v_axis = self.wl.images[idx].variable_axis
[…]
