### 2023-11-21 39d623d
reset mean kv value
