### 2023-11-21 e06d1d5
set same window for low density
