### 2023-11-21 abcebde
refactor to single use disklocator for MTMF
