### 2023-11-21 524ae35
allow to pass down detection conditions
