+++ tests_basic/test_winstonlutz.py
@@ -15,6 +15,7 @@ from pylinac import WinstonLutz, WinstonLutzMultiTargetMultiField
 from pylinac.core.array_utils import create_dicom_files_from_3d_array
 from pylinac.core.geometry import Vector, vector_is_close
 from pylinac.core.io import TemporaryZipDirectory
+from pylinac.core.metrics import is_round
 from pylinac.core.scale import MachineScale
 from pylinac.winston_lutz import (
     Axis,
@@ -654,6 +655,16 @@ class GeneralTests(TestCase):
     def test_plot_bb_location(self):
         # shouldn't raise
         self.wl.plot_location()
+
+    def test_detection_conditions_passed_down(self):
+        class WL(WinstonLutz):
+            # override the default detection conditions
+            detection_conditions = [is_round]
+
+        wl = WL.from_demo_images()
+        self.assertEqual(len(wl.detection_conditions), 1)
+        self.assertEqual(wl.detection_conditions[0], is_round)
+        self.assertEqual(wl.detection_conditions, wl.images[0].detection_conditions)
 
 
 class TestPublishPDF(TestCase):
