### 2023-11-22 c70bcdb
try minimizing memory usage
