### 2023-11-22 9191f16
do hash check vs copy
