### 2023-11-22 a9513c1
refactor boundary detection into function
