### 2023-11-22 a2133d3
update docstrings and docs
