### 2023-11-22 fe1df0b
refactor the field detector metric to use linear thresholding vs percentiles.
