### 2023-11-27 #6001 @jrkerns
Hi @Kevin-Mattheus-Moerman,

Thanks for taking the time to look this over. I have edited the zenodo listing to reflect the first change (to my understanding): https://zenodo.org/records/10145069. I adjusted the paper to reflect the last 3 changes: https://github.com/jrkerns/pylinac/commit/35c94407ba64dcb8d77953024a66e333da39a73c

Let me know if there's anything else that needs adjusting. Thanks!
