### 2023-11-27 e169da0
refactor to general method
