+++ pylinac/core/metrics.py
@@ -115,9 +115,13 @@ def is_right_area_square(region: RegionProperties, *args, **kwargs) -> bool:
     return low_bound_expected_area < field_area < high_bound_expected_area
 
 
-def deduplicate_points(
-    original_points: list[Point], new_points: list[Point], min_separation_px
-) -> list[Point]:
+def deduplicate_points_and_boundaries(
+    original_points: list[Point],
+    new_points: list[Point],
+    min_separation_px: float,
+    original_boundaries: list[np.ndarray],
+    new_boundaries: list[np.ndarray],
+) -> (list[Point], list[np.ndarray]):
     """Deduplicate points that are too close together. The original points should be the
     starting point. The new point's x, y, and z values are compared to the existing points.
     If the new point is too close to the original point, it's dropped. If it's sufficiently
@@ -126,13 +130,15 @@ def deduplicate_points(
     We assume the original points are already deduplicated. When used in a loop starting from an empty list
     this is true."""
     combined_points = original_points
-    for new_point in new_points:
+    combined_boundaries = original_boundaries
+    for new_point, new_boundary in zip(new_points, new_boundaries):
         for original_point in original_points:
             if new_point.distance_to(original_point) < min_separation_px:
                 break
         else:
             combined_points.append(new_point)
-    return combined_points
+            combined_boundaries.append(new_boundary)
+    return combined_points, combined_boundaries
 
 
 class MetricBase(ABC):
@@ -189,7 +195,175 @@ class MetricBase(ABC):
         pass
 
 
-class DiskRegion(MetricBase):
+class GlobalDiskLocator(MetricBase):
+    name: str
+    points: list[Point]
+    y_boundaries: list[np.ndarray] = []
+    x_boundaries: list[np.ndarray] = []
+
+    def __init__(
+        self,
+        radius_mm: float,
+        radius_tolerance_mm: float,
+        detection_conditions: list[Callable[[RegionProperties, ...], bool]] = (
+            is_round,
+            is_right_size_bb,
+            is_right_circumference,
+        ),
+        min_number: int = 1,
+        max_number: int | None = None,
+        min_separation_mm: float = 5,
+        name="Global Disk Locator",
+    ):
+        """Finds BBs globally within an image.
+
+        Parameters
+        ----------
+        radius_mm : float
+            The radius of the BB in mm.
+        radius_tolerance_mm : float
+            The tolerance of the BB radius in mm.
+        detection_conditions : list[callable]
+            A list of functions that take a regionprops object and return a boolean.
+            The functions should be used to determine whether the regionprops object
+            is a BB.
+        min_number : int
+            The minimum number of BBs to find. If not found, an error is raised.
+        max_number : int, None
+            The maximum number of BBs to find. If None, no maximum is set.
+        min_separation_mm : float
+            The minimum distance between BBs in mm. If BBs are found that are closer than this,
+            they are deduplicated.
+        name : str
+            The name of the metric.
+        """
+        self.radius = radius_mm
+        self.radius_tolerance = radius_tolerance_mm
+        self.detection_conditions = detection_conditions
+        self.name = name
+        self.min_number = min_number
+        self.max_number = max_number or 1e3
+        self.min_separation_mm = min_separation_mm
+
+    def _calculate_sample(
+        self, sample: np.ndarray, top_offset: int, left_offset: int
+    ) -> (list[Point], list[np.ndarray], list[RegionProperties]):
+        """Find up to N BBs/disks in the image. This will look for BBs at every percentile range.
+        Multiple BBs may be found at different threshold levels."""
+
+        # The implementation difference here from the original isn't large,
+        # But we need to detect MULTIPLE bbs instead of just one.
+        bbs = []
+        boundaries = []
[…]
