+++ tests_basic/core/test_profile_metrics.py
@@ -1,5 +1,6 @@
 from unittest import TestCase
 
+import numpy as np
 from pydicom import Dataset
 
 from pylinac.core.geometry import Point
@@ -11,9 +12,10 @@ from pylinac.core.image_generator.layers import (
     Layer,
 )
 from pylinac.core.metrics import (
-    DiskLocator,
-    GlobalDiskLocator,
+    GlobalFieldLocator,
+    GlobalSizedDiskLocator,
     GlobalSizedFieldLocator,
+    SizedDiskLocator,
     deduplicate_points_and_boundaries,
 )
 from tests_basic.utils import get_file_from_cloud_test_repo
@@ -60,14 +62,17 @@ def create_open_field_image(
 def create_multi_open_field(
     field_sizes=((50, 50), (50, 50)),
     offsets=((0, 0), (40, 40)),
+    alphas: tuple[float] | None = None,
     field: type[Layer] = FilteredFieldLayer,
 ) -> Dataset:
     as1000 = AS1000Image(
         sid=1000
     )  # this will set the pixel size and shape automatically
-    for field_size, offset in zip(field_sizes, offsets):
+    if alphas is None:
+        alphas = len(field_sizes) * [1]
+    for field_size, offset, alpha in zip(field_sizes, offsets, alphas):
         as1000.add_layer(
-            field(field_size_mm=field_size, cax_offset_mm=offset)
+            field(field_size_mm=field_size, cax_offset_mm=offset, alpha=alpha)
         )  # create a 50x50mm square field
     as1000.add_layer(
         GaussianFilterLayer(sigma_mm=2)
@@ -79,7 +84,7 @@ def create_multi_open_field(
 class TestGeneralMetric(TestCase):
     def test_modifying_image_fails(self):
         # user should not modify the image under the hood
-        class DiskModifier(DiskLocator):
+        class DiskModifier(SizedDiskLocator):
             def calculate(self) -> Point:
                 calc = super().calculate()
                 self.image.crop(10)  # modifying the image
@@ -110,23 +115,23 @@ class TestGlobalDiskLocator(TestCase):
 
     def test_isocal_image_default(self):
         bbs = self.xim.compute(
-            metrics=GlobalDiskLocator(radius_mm=3.5, radius_tolerance_mm=1)
+            metrics=GlobalSizedDiskLocator(radius_mm=3.5, radius_tolerance_mm=1)
         )
-        self.assertEqual(len(bbs), 15)
+        self.assertEqual(len(bbs), 16)
 
     def test_min_number_fails(self):
         # we should get a failure if the min number isn't found
         with self.assertRaises(ValueError):
             self.xim.compute(
-                metrics=GlobalDiskLocator(
-                    radius_mm=3.5, radius_tolerance_mm=1, min_number=16
+                metrics=GlobalSizedDiskLocator(
+                    radius_mm=3.5, radius_tolerance_mm=1, min_number=17
                 )
             )
 
     def test_max_number_is_close(self):
         # the max number is soft, but we can stop early if we find at least the max number
         bbs = self.xim.compute(
-            metrics=GlobalDiskLocator(
+            metrics=GlobalSizedDiskLocator(
                 radius_mm=3.5, radius_tolerance_mm=1, max_number=5
             )
         )
@@ -136,22 +141,32 @@ class TestGlobalDiskLocator(TestCase):
         # set the radius too high to test that no bbs are found
         with self.assertRaises(ValueError):
             self.xim.compute(
-                metrics=GlobalDiskLocator(radius_mm=10, radius_tolerance_mm=1)
+                metrics=GlobalSizedDiskLocator(radius_mm=10, radius_tolerance_mm=1)
             )
 
     def test_min_separation(self):
[…]
