+++ tests_basic/core/test_profile_metrics.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from unittest import TestCase
 
 import numpy as np
