### 2023-11-29 dab1a90
remove demo file from PF multiple image loader
