+++ pylinac/metrics/image.py
@@ -1,0 +1,713 @@
+from __future__ import annotations
+
+import math
+import typing
+from abc import ABC, abstractmethod
+from typing import Any, Callable
+
+import numpy as np
+from matplotlib import pyplot as plt
+from skimage import measure, segmentation
+from skimage.measure._regionprops import RegionProperties
+
+from ..core.array_utils import invert, stretch
+from ..core.geometry import Point
+from ..metrics.features import (
+    is_right_area_square,
+    is_right_circumference,
+    is_right_size_bb,
+    is_right_square_perimeter,
+    is_round,
+    is_solid,
+    is_symmetric,
+)
+from ..metrics.utils import deduplicate_points_and_boundaries, get_boundary
+
+if typing.TYPE_CHECKING:
+    from ..core.image import BaseImage
+
+
+class MetricBase(ABC):
+    """Base class for any 2D metric. This class is abstract and should not be instantiated.
+
+    The subclass should implement the ``calculate`` method and the ``name`` attribute.
+
+    As a best practice, the ``image_compatibility`` attribute should be set to a list of image classes that the metric
+    is compatible with. Image types that are not in the list will raise an error. This allows
+    compatibility to be explicit. However, by default this is None and no compatibility checking is done.
+    """
+
+    unit: str = ""
+    image: BaseImage
+    image_compatibility: list[BaseImage] | None = None
+    name: str
+
+    def inject_image(self, image: BaseImage):
+        """Inject the image into the metric."""
+        if self.image_compatibility is not None and not isinstance(
+            image, self.image_compatibility
+        ):
+            raise TypeError(f"Image must be one of {self.image_compatibility}")
+        self.image = image
+
+    def context_calculate(self) -> Any:
+        """Calculate the metric. This also checks the image hash to attempt to ensure no changes were made."""
+        img_hash = hash(self.image.array.tobytes())
+        calculation = self.calculate()
+        # check no modifications
+        if hash(self.image.array.tobytes()) != img_hash:
+            raise RuntimeError(
+                "A metric modified an image. This is not allowed as this could affect other, downstream metrics. Change"
+                "the calculate method to not modify the underlying image."
+            )
+        return calculation
+
+    @abstractmethod
+    def calculate(self) -> Any:
+        """Calculate the metric. Can return anything"""
+        pass
+
+    def plot(self, axis: plt.Axes, **kwargs) -> None:
+        """Plot the metric"""
+        pass
+
+    def additional_plots(self) -> list[plt.figure]:
+        """Plot additional information on a separate figure as needed.
+
+        This should NOT show the figure. The figure will be shown
+        via the ``metric_plots`` method. Calling show here would
+        block other metrics from plotting their own separate metrics.
+        """
+        pass
+
+
+class GlobalSizedDiskLocator(MetricBase):
+    name: str
+    points: list[Point]
+    y_boundaries: list[np.ndarray]
+    x_boundaries: list[np.ndarray]
+
+    def __init__(
+        self,
+        radius_mm: float,
+        radius_tolerance_mm: float,
+        detection_conditions: list[Callable[[RegionProperties, ...], bool]] = (
+            is_round,
+            is_right_size_bb,
+            is_right_circumference,
+        ),
+        min_number: int = 1,
+        max_number: int | None = None,
+        min_separation_mm: float = 5,
+        name="Global Disk Locator",
+    ):
+        """Finds BBs globally within an image.
+
+        Parameters
+        ----------
+        radius_mm : float
+            The radius of the BB in mm.
+        radius_tolerance_mm : float
+            The tolerance of the BB radius in mm.
+        detection_conditions : list[callable]
+            A list of functions that take a regionprops object and return a boolean.
[…]
