+++ docs/source/topics/image_metrics.rst
@@ -1,7 +1,7 @@
 .. _image-metrics:
 
-Custom Image Metrics
-====================
+Images & 2D Metrics
+===================
 
 .. versionadded:: 3.16
 
@@ -21,23 +21,23 @@ Use Cases
 Tool Legend
 -----------
 
-+---------------------------------------------------------+-----------------------------------------------------+--------------------------------------------------------------------+
-| Use Case                                                | Constraint                                          | Class                                                              |
-+=========================================================+=====================================================+====================================================================+
-| Find the location of a BB in the image                  | The BB size and location is known approximately     | :class:`~pylinac.core.metrics.SizedDiskLocator`                    |
-+---------------------------------------------------------+-----------------------------------------------------+--------------------------------------------------------------------+
-| Find the ROI properties of a BB in the image            | The BB size and location is known approximately     | :class:`~pylinac.core.metrics.SizedDiskRegion`                     |
-+---------------------------------------------------------+-----------------------------------------------------+--------------------------------------------------------------------+
-| Find the location of *N* BBs in the image               | The BB size is known approximately                  | :class:`~pylinac.core.metrics.GlobalSizedDiskLocator`              |
-+---------------------------------------------------------+-----------------------------------------------------+--------------------------------------------------------------------+
-| Find the location of a square field in an image         | The field size is known approximately               | :class:`~pylinac.core.metrics.GlobalSizedFieldLocator`             |
-+---------------------------------------------------------+-----------------------------------------------------+--------------------------------------------------------------------+
-| Find the locations of *N* square fields in an image     | The field size is not known                         | :class:`~pylinac.core.metrics.GlobalFieldLocator`                  |
-+---------------------------------------------------------+-----------------------------------------------------+--------------------------------------------------------------------+
-| Find the location of a circular field in an image       | The field size and location are known approximately | :class:`~pylinac.core.metrics.SizedDiskLocator` (``invert=False``) |
-+---------------------------------------------------------+-----------------------------------------------------+--------------------------------------------------------------------+
-| Find the ROI properties of a circular field in an image | The field size and location are known approximately | :class:`~pylinac.core.metrics.SizedDiskRegion` (``invert=False``)  |
-+---------------------------------------------------------+-----------------------------------------------------+--------------------------------------------------------------------+
++---------------------------------------------------------+-----------------------------------------------------+---------------------------------------------------------------------+
+| Use Case                                                | Constraint                                          | Class                                                               |
++=========================================================+=====================================================+=====================================================================+
+| Find the location of a BB in the image                  | The BB size and location is known approximately     | :class:`~pylinac.metrics.image.SizedDiskLocator`                    |
[…]
