### 2023-11-30 1cf7a78
update readme and fix various doc build errors
