+++ README.rst
@@ -29,6 +29,13 @@
     :alt: GitHub commit activity (branch)
 
 
+.. contents::
+    :class: this-will-duplicate-information-and-it-is-still-useful-here
+
+
+Introduction
+------------
+
 Pylinac provides TG-142 quality assurance (QA) tools to Python programmers in the field of
 therapy and diagnostic medical physics.
 
@@ -38,19 +45,13 @@ Most scripts can be utilized with less than 10 lines of code.
 The library also contains lower-level `modules & tools <http://pylinac.readthedocs.org/en/latest/core_modules.html>`_
 for creating your own image analysis algorithms.
 
-The major features of the entire package include:
+The major attributes of the package are:
 
 * Simple, concise image analysis API
 * Automatic analysis of imaging and performance metrics like MTF, Contrast, ROIs, etc.
 * PDF report generation for solid documentation
 * Automatic phantom registration even if you don't set up your phantom perfect
 * Image loading from file, ZIP archives, or URLs
-
-Documentation
--------------
-
-To get started, install the package, run the demos, view the API docs, and learn the module design, visit the
-`Full Documentation <http://pylinac.readthedocs.org/>`_ on Read The Docs.
 
 Installation
 ------------
@@ -63,322 +64,570 @@ Install via ``pip``:
 
 See the `Installation page <https://pylinac.readthedocs.io/en/latest/installation.html>`_ for further details.
 
-Quick Start Guide
------------------
+Documentation
+-------------
 
-Below are the high-level tools currently available:
+To get started, install the package, run the demos, view the API docs, and learn the module design, visit the
+`Full Documentation <http://pylinac.readthedocs.org/>`_ on Read The Docs.
 
-* `TG-51 & TRS-398 Absolute Dose Calibration <http://pylinac.readthedocs.org/en/latest/calibration_docs.html>`__ -
-   Input the raw data and pylinac can calculate either individual values (kQ, PDDx, Pion, etc) or use the
-   provided classes to input all measurement data and have it calculate all factors and dose values automatically.
+Features
+--------
 
-   Example script:
+Low-level Tooling
+~~~~~~~~~~~~~~~~~
 
-   .. code-block:: python
+DICOM, XIM & Image Loading
+##########################
 
-            from pylinac import tg51, trs398
+Load DICOM files, XIM, and TIFF images:
 
-            ENERGY = 6
-            TEMP = 22.1
-            PRESS = tg51.mmHg2kPa(755.0)
-            CHAMBER = "30013"  # PTW
-            P_ELEC = 1.000
-            ND_w = 5.443  # Gy/nC
-            MU = 200
-            CLINICAL_PDD = 66.5
+.. code-block:: python
 
-            tg51_6x = tg51.TG51Photon(
-                unit="TrueBeam1",
-                chamber=CHAMBER,
-                temp=TEMP,
-                press=PRESS,
-                n_dw=ND_w,
-                p_elec=P_ELEC,
-                measured_pdd10=66.4,
-                lead_foil=None,
-                clinical_pdd10=66.5,
-                energy=ENERGY,
-                voltage_reference=-300,
-                voltage_reduced=-150,
-                m_reference=(25.65, 25.66, 25.65),
-                m_opposite=(25.64, 25.65, 25.65),
-                m_reduced=(25.64, 25.63, 25.63),
-                mu=MU,
-                tissue_correction=1.0,
-            )
+  from pylinac import image
+  from pylinac.core.image import XIM
 
-            # Done!
-            print(tg51_6x.dose_mu_dmax)
+  my_dcm = image.load("path/to/my/image.dcm")
[…]
