### 2023-12-02 #6001 @Kevin-Mattheus-Moerman
@jrkerns thanks for making those changes. The version tag is now correctly set for the archive. One remaining issue is the title for the archive, it should match the paper title `Pylinac: Image analysis for routine quality assurance in radiotherapy`. Can you please edit the archive to do this?
